+++ Editor/Drawers/FolderDrawers/AssetsFolderDrawer/AssetsFolderAttributeDrawerIMGUI.cs
@@ -7,13 +7,13 @@ using SaintsField.Editor.Utils;
 using SaintsField.Interfaces;
 using UnityEditor;
 using UnityEngine;
-using Object = UnityEngine.Object;
 
 namespace SaintsField.Editor.Drawers.FolderDrawers.AssetsFolderDrawer
 {
-    public partial class AssetFolderAttributeDrawer
+    public partial class AssetsFolderAttributeDrawer
     {
         private static Texture2D _folderIcon;
+        private static Texture2D _openIcon;
 
         private sealed class InfoIMGUI
         {
@@ -48,7 +48,7 @@ namespace SaintsField.Editor.Drawers.FolderDrawers.AssetsFolderDrawer
 
         protected override float GetPostFieldWidth(Rect position, SerializedProperty property, GUIContent label,
             IReadOnlyList<PropertyAttribute> allAttributes, ISaintsAttribute saintsAttribute, int index,
-            FieldInfo info, object parent) => SingleLineHeight;
+            FieldInfo info, object parent) => SingleLineHeight * 2;
 
         protected override bool DrawPostFieldImGui(Rect position, Rect fullRect, SerializedProperty property,
             GUIContent label, ISaintsAttribute saintsAttribute, int index,
@@ -58,16 +58,31 @@ namespace SaintsField.Editor.Drawers.FolderDrawers.AssetsFolderDrawer
             UpdateStatus(cache, property);
 
             _folderIcon ??= Util.LoadResource<Texture2D>("folder.png");
-            Rect buttonRect = new Rect(position)
+            _openIcon ??= Util.LoadResource<Texture2D>("open-ext.png");
+            Rect linkButtonRect = new Rect(position)
             {
                 x = position.x + 1,
-                width = position.width - 2,
+                width = SingleLineHeight - 1,
+            };
+            Rect pickButtonRect = new Rect(position)
+            {
+                x = position.x + SingleLineHeight,
+                width = SingleLineHeight - 1,
             };
 
             if (property.propertyType == SerializedPropertyType.String)
             {
+                Object folderObj = GetFolderObjectIMGUI(property.stringValue);
+                using (new EditorGUI.DisabledScope(folderObj == null))
+                {
+                    if (GUI.Button(linkButtonRect, _openIcon, GUIStyle.none))
+                    {
+                        EditorGUIUtility.PingObject(folderObj);
+                    }
+                }
+
                 FolderAttribute folderAttribute = (FolderAttribute)saintsAttribute;
-                if (GUI.Button(buttonRect, _folderIcon, GUIStyle.none))
+                if (GUI.Button(pickButtonRect, _folderIcon, GUIStyle.none))
                 {
                     (string error, string actualFolder) = OnClick(property, folderAttribute);
                     if (error == "")
@@ -97,7 +112,8 @@ namespace SaintsField.Editor.Drawers.FolderDrawers.AssetsFolderDrawer
             {
                 using (new EditorGUI.DisabledScope(true))
                 {
-                    GUI.Button(buttonRect, _folderIcon, GUIStyle.none);
+                    GUI.Button(linkButtonRect, _openIcon, GUIStyle.none);
+                    GUI.Button(pickButtonRect, _folderIcon, GUIStyle.none);
                 }
             }
 
@@ -205,6 +221,11 @@ namespace SaintsField.Editor.Drawers.FolderDrawers.AssetsFolderDrawer
             }
 
             return "";
+        }
+
+        private static Object GetFolderObjectIMGUI(string value)
+        {
+            return Directory.Exists(value) ? AssetDatabase.LoadAssetAtPath<Object>(value) : null;
         }
     }
 }