+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 5.29.4 ##
+
+1.  Fix: `SceneReference` darg fix [#445](https://github.com/TylerTemp/SaintsField/issues/445)
+2.  Fix: `SceneReference` now refreshes when scenes in Build Settings change
+3.  Rename: `AssetFolder` -> `AssetsFolder`; `ResourceFolder` -> `ResourcesFolder` (Old name is kept but marked as obsolete)
+4.  Add: `AssetsFolder` and `ResourcesFolder` now have a button to locate the selected folder in the Project window.
+
 ## 5.29.3 ##
 
 1.  Add: `LeftToggle` now works with `ShowInInspector` & `Button`